Add Lightbox view
Create a Lightbox view which listens to the document for clicks on
elements with an externally-specified attribute, and matches that up
with a supplied list of image data.

Currently this listens to the document continuously for escape presses
and document clicks, but it could be easily adapted to allow for more
fine-grained control of listener addition/removal if there was some
kind of significant performance advantage in a specific context.